+++ Delta.CertXplorer/src/Delta.CertXplorer.Common/DocumentModel/DefaultDocumentHandler.cs
@@ -6,11 +6,16 @@ namespace Delta.CertXplorer.DocumentModel
 {
     internal class DefaultDocumentHandler : BaseDocumentHandler<DefaultDocumentView>
     {
+        public override string HandlerName
+        {
+            get { return "Default Document Handler [Embedded]"; }
+        }
+
         protected override bool CanHandleSource(IDocumentSource source)
         {
             return (source != null && source is FileDocumentSource);
         }
-
+        
         protected override IDocument CreateDocumentFromSource()
         {
             var data = File.ReadAllBytes(Source.Uri);