+++ Delta.CertXplorer/src/Delta.CertXplorer.Common/DocumentModel/DocumentFactory.cs
@@ -1,4 +1,8 @@
-﻿namespace Delta.CertXplorer.DocumentModel
+﻿using System.Collections.Generic;
+using System.Windows.Forms;
+using Delta.CertXplorer.UI;
+
+namespace Delta.CertXplorer.DocumentModel
 {
     /// <summary>
     /// Allows for creation of a document manager service.
@@ -14,7 +18,7 @@
         /// </returns>
         public static IDocumentManagerService CreateDocumentManagerService(IDocumentBasedUI chrome)
         {
-            return new DocumentManagerService(chrome);
+            return new DocumentManagerService(chrome, ChooseDocumentHandler);
         }
 
         /// <summary>
@@ -26,6 +30,18 @@
         public static IDocumentHandlerRegistryService CreateDocumentHandlerRegistryService()
         {
             return new DocumentHandlerRegistryService();
+        }
+
+        private static IDocumentHandler ChooseDocumentHandler(IDocumentBasedUI ui, IEnumerable<IDocumentHandler> handlers)
+        {
+            using (var dialog = new ChooseDocumentHandlerDialog())
+            {
+                dialog.SetDocumentHandlers(handlers);
+                if (dialog.ShowDialog() == DialogResult.OK)
+                    return dialog.SelectedDocumentHandler;
+            }
+
+            return null;
         }
     }
 }