+++ Delta.CertXplorer/src/Delta.CertXplorer.Common/DocumentModel/IDocumentHandler.cs
@@ -9,5 +9,7 @@
         IDocument Document { get; }
         
         IDocumentView CreateView();
+
+        string HandlerName { get; }
     }
 }